ose-python/requirements.txt
flask==2.3.2
sqlalchemy==2.0.20
redis==4.6.0
prql-python==0.9.0
python-dotenv==1.0.0
pymysql==1.1.1
pytest==7.4.0
apscheduler==3.11.0
sqlparse==0.5.3
